requirements.txt
streamlit
streamlit-mic-recorder
langchain
langchain-community
openai
python-dotenv
pyperclip
streamlit-extras
streamlit-lottie
requests
streamlit-option-menu
python-docx
react
streamlit-component-lib
fpdf